refactor(checker): split type-parameter helpers out of type_environment/core.rs
Extract the type-parameter extraction/resolution/required-count cluster
(get_type_params_for_symbol, get_display_type_params_for_symbol,
count_required_type_params[_from_ast], extract_type_params_from_decl, the
jsdoc-template helpers, type_param_names_in_arena, etc.) from
crates/tsz-checker/src/state/type_environment/core.rs into a new sibling
submodule type_environment/type_params.rs. Pure code move of one self-contained
cluster of CheckerState methods, no logic change: core.rs drops 2537 -> 1461
LOC, back under the 2000-LOC checker boundary, so it is removed from the
arch_guard LINE_LIMIT_CHECKS allowlist (8 -> 7 oversized checker files) and its
FILE_LINE_LIMIT ratchet tightens 2568 -> 1463.

Verified: the cluster's private helpers are only called within the cluster, and
no moved method forward-references a private helper left in core.rs or calls a
module-private item (the ALLOW_CONCRETE_REMAPPED_KEY_FALLBACK thread-local stays
in core.rs and is not referenced by the moved code). tsz-checker builds clean
with no new warnings.

AgentName: M1-Opus
