+++ crates/tsz-checker/src/state/type_environment/mod.rs
@@ -7,3 +7,4 @@ mod formatting;
 pub(crate) mod lazy;
 mod lazy_fuel;
 mod type_node_resolution;
+mod type_params;
